backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
blinker==1.7.0
cachetools==6.1.0
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.6
colorama==0.4.6
cryptography==41.0.7
cssselect==1.3.0
cssutils==2.11.1
django==4.2.11
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.6
jinja2==3.1.2
kombu==5.3.5
lxml==5.4.0
markdown==3.4.1
markupsafe==2.1.5
more-itertools==10.7.0
mysqlclient==2.2.7
packaging==24.0
pillow==11.2.1
premailer==3.10.0
psutil==5.9.8
pycparser==2.22
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
requests==2.31.0
setuptools==80.9.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.10.0
urllib3==2.0.7
vine==5.1.0
whitenoise==6.5.0
django-cors-headers==4.7.0
